[REFACTOR] global dto 패키지 분리
